requirements.txt
pandas
pinecone
groq
streamlit
sentence-transformers
python-dotenv
pytest
spotipy
torchvision
